Extract examples
